Wrap client certificate errors in EasResponse.
Client certificate errors and bad username/error passwords just return
403 and it's difficult for us to differentiate the two. The easiest way
right now is to track it using a dummy KeyManager that can detect when a
certificate is requested upon connection establishment (way low in the
bowels of the Apache HTTP stack).

This change propagates that information up to the EasResponse and
encapsulates it there.

In the future, we should be more flexible as there can theoretically be
servers that REQUEST certificates, but do not REQUIRE it, but this code
will present an error regardless.

Change-Id: I7ee36e2c2ab06bdb8ce34b8967b7cb241812ac96