Exceptions to recurring events are not deleted
If an exception from a recurring event is created
from the server and then the whole event serie is
deleted from the phone client the exceptions will be left
in the calendar database.
This is because it is up to the phone client to delete
the events that is marked to be deleted. No delete
commands will be returned by the exchange server to
the client that orders the deletion of the recurring series.

The problem is that the phone client only deletes the main
recurring event and does nothing to the exceptions, which
lead to ghost events in the calendar database. This fix makes
sure that all events related to the main event are deleted.

Change-Id: Ief5f693cc5d05b28df42a26020c0997a344a5f35