Unable to sync email after password expired twice
When the device password expire on an account with the "Enforce
password history" enabled, the device will delete the mailbox data from
the database and update the user data from the server. After the account
expires a second time, the device will still use the cached account user
data that was fetched from the first mailbox reset. This fix will
disregard the cached user data in the event of a reload request of the
folder list (the password has expired) and query the database for the
correct user data.

Change-Id: I2e6e0bab08cf7cd3bb9c880bb5d7cb7b3f61c614