Fix account sync.
We no longer specify sync intervals per mailbox. Instead,
the account sync interval is always used, and each mailbox
simply says whether it participates in that sync. This
allows us to change "account sync" to mean that we sync all
mailboxes that we are actively syncing (and push uses the
same definition of which mailboxes to check).

Also, this change implements "account only" sync.

Change-Id: I528ef5f9f0b857fd2a31a95068ef1b5ff62a5a56