Decouple operation management from EasServerConnection.
This restructuring touches a regrettable number of files,
but allows us to put common code in a common base class.
(Before, it was messy because it would be muddled with the
connection management code.) Specifically, we want common
error paths to be handled in one place, although this isn't
implemented yet.

The new class of interest is EasOperation. This CL changes
Ping to use this base class. Future changes will switch all
operations to work this way.

Change-Id: I1bd26336e8916cafe592352f8ee05616bce8181c