Stop the SyncAdapterService if no accounts
b/10360861
If the service got into a running state, even if there
were no exchange accounts on the device, it would
automatically get restarted and never stopped.
Now, even upon getting started, if there are no accounts,
it will stop itself.
It still automatically gets started up when the user
enters the app, this is due to some code that processes
services.xml to start up any services. That's okay for
now, at least the service will get immediately destroyed
if it's not needed.

Change-Id: I7f62443fc5fa1e0a46cfbbea8d7b6792cdf0a6cf