Handle IoException Properly
Report IoException to Sync Manager as an Io Exception instead of a Parse Exception
so that it get's retried after a backoff.

I was able to reproduce the symptoms int he attached bug by faking a throw in the
Folder Sync operation.

The sync would never complete unless started manually.

I then tested the fix the same way and verified that the failure resulted in a new
sync scheduled with an exponential backoff.

Bug: 11047819
Change-Id: I18daf235bfce208ace1d5a44b6a2a420fe3969bf