Exchange: Skip the tag "TRUNCATED" when parse the result of fetching entire mail
When we send the request to get the whole mail. In most instances, the reponse will not
contains the "TRUNCATED" tag, but sometimes it will contain the "TRUNCATED" tag such as
the hotmail server. And caused by we request to fetch the whole mail and the truncated
must be false. So we didn't care about the value, and could skip this tag.

CRs-fixed: 543194

Change-Id: Ibd34464baa1abe7e7a48a2e3dcdcf6ff81d7face