Properly manage Sync Window
Start with 1
Increment whenever we detect a possible loop
Reset back to one whenever we do not detect a possible loop.

Not sure how this worked before. We probably get SYNC_RESULT_MORE_AVAILABLE
and either increment the window to 1 or reset it 1. But on this particular
server, maybe WindowSize=0 does not result in a SYNC_RESULT_MORE_AVAILABLE.

Bug: 11247743
Change-Id: Ic25f694288242d4dc6630739d53a39ba1a3a6a9f