Backoff if sync of a folder fails
b/11636563
We still don't yet know why this particular command is
getting a protocol error, but we at least should not instantly
ping again, as this puts us in a rapid loop.
Now, if sync on a folder fails for any reason, we will not
immediately restart the ping for that folder's account. We'll
set an alarm to wait for 5 minutes before starting the ping
again. It makes sense to wait, if a sync fails, whether due
to an IO error, authentication error, transient server error,
it likely won't succeed only a half second later. It's good
to back off a bit.
At some later time maybe we should add a more advanced backoff
policy, but for now holding off for 5 minutes will drastically
reduce battery drain with a fairly low risk of introducing
really bad new problems.

Change-Id: I909dfcb8e6f4e10991cf15a2c08daab83dae97e7