Improve birthday handling in contacts upsync.
Before this change, we assumed the string in the contacts
provider is in the proper format for Exchange. However, this
is not always true, so we now try to parse it using two
possible formats, and then make sure to format it correctly
when forming the request.

If the birthday isn't in one of the two known formats, we
do not send it in the upsync. This is bad, but will not arise
with the default contacts app.

Bug: 11636563
Change-Id: Ib7268409f5c3f39ecb345a2628ed66b73c40604c