Do not send extraneous tags when confirming a remote wipe. b/12921830
This is an integration of an Exchange patch from Motorola.
getRequestEntity() was modified to not add tags that were
not part of the protocol spec for remote wipe ack.
Another big change was to unwrap this function so it could
be unit tested w/o the need for bigger, bulkier classes.
This lead to the creation of populateRequestEntitySerializer()
and changes to other functions to make then reentrant and
potentially static if possible. Oh...fixed a typo too.

Change-Id: I0c9167c9609b3d6a6f77f3f3118d9ae790077b2d
(cherry picked from commit 60b69ecd47d872c845b6bca46e4937d505a78101)