Make sending an EasOperation
I needed to make some changes to the EasOperation to
allow additional functionality to be put in the child
classes:
* Now child classes can override handleHttpError to
change the result code based upon the http response.
* Child classes can override onRequestMade to do
any cleanup that must be done. This is needed because
when sending a message, we need to put the message info
into a temporary file. Only after sending the httpEntity
can we delete that file.

Change-Id: I5554393d521116ce28e396a90c8ddffd48b9a3d3