b/13349278. Fix multiple downloads of the same attachment.
The problem was due to a different set of error codes being used
internally within the EasLoadAttachment class and the externally
used EmailServiceStatus codes.  The fix was to be consistent about
using interal codes only internally and using the EmailServiceStatus
codes when pushing status (via callbacks) back to the
AttachmentDownload service.

Change-Id: I8b7705384c156d8e4a814059bb7b2ecbbb57d574