Exchange: Fix NPE caused by the nonexistent account
During the action of syncing in onPerformSync(), if the querying
account is deleted, content resolver will return a null cursor.
When onPerformSync() continues to access this null cursor without
check, NPE is thrown.

Need to check the cursor. If it is null, there will be nothing to
sync. So stop the performance and return.

CRs-fixed: 655249

Change-Id: I857d800266a174ff1e94ab6ee2f959dcc9e5f258