Update IEmailService
This is not yet finished, but it's needed to fix push syncs on
exchange, which are currently broken.
Now the actual work of syncing up and down is done in EasService.
EmailSyncAdapterService just gets the onPerformSync() call and
has the EasService do the work. This makes way for eventually
moving the sync adapter and the ping synchronizer out of exchange
and into email.

Change-Id: Ic5608887230393560612c9fa45992e4413317bfd