Make sure that we have an EasService in onPerformSync
The call to ServiceConnection.onServiceConneted is asynchronous
from the call to bindService. It's therefor possible for
onPerformSync to be called before we have the service.
Now, in onPerformSync(), if the service isn't yet present,
we will wait for up to 10 seconds for it to appear.
If it still is not present after that, the sync will fail.
I'm not sure what else we could do at that point, there's
no good reason why it should take 10 seconds for another service
in our own app to connect.

Change-Id: I66de8feedf0373cfebc93b23d84f83642af1a78f