Correctly set IS_PRIMARY in CalendarUtilities
b/16847896
There is a preexisting bug here, which is that, as far as I
can tell, changes to an existing calendar will not actually
be propagated to the calendar app. Only new calendars are
correctly updated. I don't know if any exchange servers actually
implement this, but the protocol allows it.
That can be fixed in a separate CL.
Bug filed: b/17647582

Change-Id: I9de7099a660e4eae9cefa7157848634767985e5d