When we migrate Exchange folders, set the sync status to INITIAL_SYNC_NEEDED
b/17443087
When we migrate exchange accounts, we copy mailboxes over to the
new account (in order to preserve sync frequency and window).
The problem with this is, you may have many accounts/mailboxes.
After starting the app, it may take quite a long time before all
mailboxes are synced. If the user visits some mailbox near the
bottom, they'll see a misleading "folder is empty" view.
Now, when migrating, we'll set the uisyncstatus to INITIAL_SYNC_NEEDED.

The issue here is that when we start a fresh sync, we temporarily
store sync info needed in RAM, clear all of the folders, then
restore the sync. I needed to add uiSyncStatus to the list of
attributes copied over.

Change-Id: I223d2e000cfd48f62ce694365cbb8ec3850a50b3