Fix search parsing
b/17231624
SearchParser delegates parsing of email contents to the
EmailSyncParser. Whenever it encounters a <Properties> tag,
it pushes that tag onto the EmailSyncParser, and then has
it parse and get message contents. It will keep doing this
until it encounters a </Properties> tag.
However, this means that that the </Properties> tag is no
longer in the input stream. The parent parser (SearchParser)
won't see it, so it won't remove it from its internal stack.
We must manually pop that from the stack or the parser is in
a confused state.

Change-Id: Ia3352fc16f8281d8fc597048df5e105ca7e6ba49