If an account does not support smart forward, then it doesn't do smart reply either
b/17720266
It turns out that if an account's FLAG_SUPPORTS_SMART_FORWARD was unset,
that did not prevent it from attempting to smart reply. We need to be
able to disable all types of smart sending, so I'm making
FLAG_SUPPORTS_SMART_FORWARD also control ability to smart reply.

In a later CL, we should probably make separate flags for these
two capabilities.

Also, add some logging for the case where we attempt a smart send, and
the server rejects it. This should be an uncommon case.

Change-Id: I0d6a54b5f7aa796b72c9eb133a84cd4099fbadcb