Exchange: fix call wrong service when reschedule an eas ping
In Lollipop stock exchange api performs ping operations through the EasService instead of EmailSyncAdapterService
Commit 85c66ca12d7debfbbcfc0a3cc3afab1673577d50 introduced an optimization to reschedule a delayed ping request
when the last ping operation ended with an error. But the scheduleDelayedPing still create the alarm against the
EmailSyncAdapterService service, which now doesn't listen for this sort of operation.
This patch changes the alarm intent to the EasService and readd the code that starts a new request ping.

Change-Id: I16784642bf04afe6c61624dc04b5d045c86155ab
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>