exchange: increase ping command reschedule time on http server error
Actually, if the Exchange Server returns with a Server Error http status for a ping command,
the ping command is continuosly execute draining the device battery and using an big amount
of data, due command asking to server every few milliseconds. Since the server usually
isn't going to recover in a short time, is better to decrease the latence of the ping
commands to a exponential amount of time (from 5 seconds to a maximum of 10 minutes).

Change-Id: Ie5e15256a1059ac10dc23983760008ee99f7959a
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>