Email: fix the mail bigger than 1M can't downloaded on Exchange account.
The exchange update the database through contentresolve, and it is IPC
it will use bundle to deliver the data. The bundle has the size limit,
so it will throw the TransactionTooLargeException, lead Exchange can't
update database, so the mail can't downloaded.

Compress the mail content before put the data to bundle, and decompress
the data in EmailProvider.

Change-Id: I66fa0fca96bb0425f1f2c56082e899eb030169ca
CRs-fixed: 943810